Added pack unpack functions for all commands

+++ interfaces/obc_gs_interface/commands/obc_gs_command_data.h
@@ -25,7 +25,7 @@ typedef enum { BOOTLOADER = 0x00, APPLICATION } programming_session_t;
 
 // CMD_DOWNLOAD_DATA
 typedef struct {
-  uint8_t *downloadData;
+  uint32_t *downloadData;
   uint8_t length;
 } download_data_cmd_data_t;
